requirements.txt
goodfire
python-dotenv
openai
pandas
numpy
matplotlib